+++ spec/db/tables.sql
@@ -1,20 +1,31 @@
-DROP TABLE IF EXISTS `simmer_test`.`notes`;
-DROP TABLE IF EXISTS `simmer_test`.`agents`;
+DROP TABLE IF EXISTS `agents`;
 
-CREATE TABLE `simmer_test`.`agents` (
+CREATE TABLE `agents` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
   `call_sign` varchar(150) NOT NULL,
-  `first` varchar(255),
-  `last` varchar(255),
+  `first` varchar(255) DEFAULT NULL,
+  `last` varchar(255) DEFAULT NULL,
   PRIMARY KEY (`id`),
-  UNIQUE INDEX (`call_sign`)
-);
+  UNIQUE KEY `call_sign` (`call_sign`)
+) ENGINE = InnoDB AUTO_INCREMENT = 3 DEFAULT CHARSET = utf8mb4;
 
-CREATE TABLE `simmer_test`.`notes` (
+DROP TABLE IF EXISTS `notes`;
+
+CREATE TABLE `notes` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
   `agent_id` int(11) NOT NULL,
-  `note` varchar(255),
+  `note` varchar(255) DEFAULT NULL,
   PRIMARY KEY (`id`),
-  INDEX (`agent_id`),
-  FOREIGN KEY (`agent_id`) REFERENCES `simmer_test`.`agents`(`id`)
-);
+  KEY `agent_id` (`agent_id`),
+  CONSTRAINT `notes_ibfk_1` FOREIGN KEY (`agent_id`) REFERENCES `agents` (`id`)
+) ENGINE = InnoDB DEFAULT CHARSET = utf8mb4;
+
+DROP TABLE IF EXISTS `table`;
+
+CREATE TABLE `table` (
+  `id` bigint(20) unsigned NOT NULL AUTO_INCREMENT,
+  `column` varchar(30) NOT NULL,
+  `count` int(11) NOT NULL,
+  PRIMARY KEY (`id`),
+  UNIQUE KEY `id` (`id`)
+) ENGINE = InnoDB AUTO_INCREMENT = 3 DEFAULT CHARSET = utf8mb4;